feat(item): extract steps to separate files for redability
