+++ components/Modal/AddItemModal.tsx
@@ -6,7 +6,7 @@ import { Button, Modal, Steps, theme, message, Spin } from 'antd';
 
 import { ITEM_CREATION_STEPS } from '../../lib/constants';
 import { useItemContext } from '../../contexts/ItemProvider';
-import { StepOne, StepTwo, StepThree, StepFour } from './steps';
+import { StepOne, StepTwo, StepThree, StepFour } from './AddItemSteps';
 import { ItemDetails } from '../../lib/types';
 
 const steps = ITEM_CREATION_STEPS.map(({ title, content }) => ({ key: title, title, content }));
