+++ components/Modal/AddItemSteps/StepOne.tsx
@@ -1,0 +1,50 @@
+import React, { useState, useEffect } from 'react';
+import {
+    Form,
+    Input,
+    InputNumber,
+    ColorPicker,
+} from 'antd';
+import { StepProps } from '../../../lib/types';
+
+const { TextArea } = Input;
+
+const StepOne = (props: StepProps) => {
+    const { updateItemDetails, item } = props;
+    const [name, setItemName] = useState<string>();
+    const [colour, setItemColour] = useState<string>();
+    const [quantity, setItemQuantity] = useState<number>();
+    const [description, setItemDescription] = useState<string>();
+    useEffect(() => {
+        updateItemDetails({
+            ...item,
+            name,
+            description,
+            quantity,
+            colour
+        })
+    }, [name, description, quantity, colour])
+    return (
+        <Form
+            layout="vertical"
+            name="generalInformationForm"
+            autoComplete="off"
+            style={{ maxWidth: 600 }}
+        >
+            <Form.Item label="Name">
+                <Input defaultValue={item?.name} onChange={e => setItemName(e?.target?.value)} />
+            </Form.Item>
+            <Form.Item label="Description">
+                <TextArea defaultValue={item?.description} rows={4} onChange={e => setItemDescription(e?.target?.value)} />
+            </Form.Item>
+            <Form.Item label="Quantity">
+                <InputNumber defaultValue={item?.quantity} onChange={e => setItemQuantity(Number(e))} />
+            </Form.Item>
+            <Form.Item label="Color">
+                <ColorPicker defaultValue={item?.colour} onChange={e => setItemColour(e?.toHexString())} />
+            </Form.Item>
+        </Form>
+    )
+};
+
+export default StepOne;
