+++ components/Modal/AddItemSteps/index.ts
@@ -1,0 +1,4 @@
+export { default as StepOne } from './StepOne';
+export { default as StepTwo } from './StepTwo';
+export { default as StepFour } from './StepFour';
+export { default as StepThree } from './StepThree';
